+++ src/main/java/roomescape/admin/console/ConsoleConfig.java
@@ -1,4 +1,4 @@
-package roomescape.controller.console;
+package roomescape.admin.console;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
