+++ src/test/java/roomescape/admin/dummy/DummyDto.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dummy;
+package roomescape.admin.dummy;
 
 import jakarta.validation.constraints.NotNull;
