+++ src/test/java/roomescape/admin/fixture/ReservationTimeApiRequestFixture.java
@@ -1,4 +1,4 @@
-package roomescape.controller.fixture;
+package roomescape.admin.fixture;
 
 import java.time.LocalTime;
 import java.util.stream.Stream;
